Added Helper Method for configuration
